requirements.txt
streamlit>=1.32.0
weaviate-client>=4.5.0
python-dotenv>=1.0.0
openai>=1.12.0
PyPDF2>=3.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
jinja2>=3.0.0
PyMuPDF==1.23.8
tqdm==4.66.1
python-docx>=0.8.11
python-pptx>=0.6.21
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pywin32>=306 